Fix missing Controller annotation. Move configs to the top-level so that component scan works properly.
