+++ src/main/resources/templates/data.html
@@ -1,10 +1,11 @@
 <!DOCTYPE html>
-<html lang="en">
+<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity3">
 <head>
-    <meta charset="UTF-8">
-    <title>Upload Your Data File</title>
+    <title>Upload your data file</title>
 </head>
 <body>
 
 </body>
 </html>
+
