+++ src/main/resources/templates/mapping.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
-<html lang="en">
+<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity3">
 <head>
-    <meta charset="UTF-8">
-    <title>Map items from your file to existing items from OpenClinica</title>
+    <title>Map items from your file to existing items from OpenClinica </title>
 </head>
 <body>
